requirements.txt
Django==4.2.24
djangorestframework==3.15.2
python-decouple==3.8
requests==2.32.3
python-telegram-bot==22.4
httpx==0.27.0
